Added view password button to single account modal. Published update to branch.

+++ AccountManager.Core/Static/ColorExtensions.cs
@@ -6,7 +6,7 @@ namespace AccountManager.Core.Static
     {
         public static Color LightenColor(this Color color, int factor)
         {
-            var lightenedColor = Color.FromArgb(5, (int)(color.R + (255 - color.R) / factor), (int)(color.G + (255 - color.G) / factor), (int)(color.B + (255 - color.B) / factor));
+            var lightenedColor = Color.FromArgb(5, (color.R + (255 - color.R) / factor), (color.G + (255 - color.G) / factor), (color.B + (255 - color.B) / factor));
             return lightenedColor;
         }
     }
